Move ChunkHandle implementation to a template base class
The ChunkHandle implementation will for the most part be shared across the classes.
It does not make sense for each class to reimplement everything.
This adds an optional template base class which can be shared.
